+++ src/roq/deribit/web_socket.cpp
@@ -150,6 +150,10 @@ void WebSocket::operator()(const core::web::Socket::Text &text) {
   parse(text.payload);
 }
 
+void WebSocket::operator()(const core::web::Socket::Binary &) {
+  log::fatal("Unexpected"_sv);
+}
+
 void WebSocket::operator()(ConnectionStatus status) {
   if (utils::update(status_, status)) {
     server::TraceInfo trace_info;